+++ graphdatascience/session/aura_api.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 import logging
 import time
 import warnings
+from collections import defaultdict
 from datetime import timedelta
 from typing import Any, Dict, List, Optional, Tuple
 from urllib.parse import urlparse
@@ -87,7 +88,7 @@ class AuraApi:
             base_uri = f"https://api-{aura_env}.neo4j-dev.io"
         return base_uri
 
-    def create_session(
+    def get_or_create_session(
         self,
         name: str,
         pwd: str,
@@ -114,7 +115,8 @@ class AuraApi:
 
         self._check_resp(response)
 
-        return SessionDetails.from_json(response.json()["data"])
+        raw_json: Dict[str, Any] = response.json()
+        return SessionDetails.from_json(raw_json["data"], raw_json.get("errors", []))
 
     def get_session(self, session_id: str) -> Optional[SessionDetails]:
         response = self._request_session.get(
@@ -126,7 +128,8 @@ class AuraApi:
 
         self._check_resp(response)
 
-        return SessionDetails.from_json(response.json()["data"])
+        raw_json: Dict[str, Any] = response.json()
+        return SessionDetails.from_json(raw_json["data"], raw_json.get("errors", []))
 
     def list_sessions(self, dbid: Optional[str] = None) -> List[SessionDetails]:
         # these are query parameters (not passed in the body)
@@ -141,7 +144,14 @@ class AuraApi:
 
         self._check_resp(response)
 
-        return [SessionDetails.from_json(s) for s in response.json()["data"]]
+        raw_json = response.json()
+
+        data: Dict[str, Any] = raw_json.get("data")
+        errors_per_session = defaultdict(list)
+        for error in raw_json.get("errors", []):
+            errors_per_session[error["id"]].append(error)
+
+        return [SessionDetails.from_json(s, errors_per_session[s["id"]]) for s in data]
 
     def wait_for_session_running(
         self,
@@ -157,6 +167,8 @@ class AuraApi:
                 return WaitResult.from_error(f"Session `{session_id}` not found -- please retry")
             elif session.status == "Ready" and session.host:  # check host needed until dns based routing
                 return WaitResult.from_connection_url(session.bolt_connection_url())
+            elif session.status == "Failed":
+                return WaitResult.from_error(f"Session `{session_id}` failed due to: {session.errors}")
             else:
                 self._logger.debug(
                     f"Session `{session_id}` is not yet running. "