+++ graphdatascience/tests/unit/test_aura_api.py
@@ -14,6 +14,7 @@ from graphdatascience.session.aura_api_responses import (
     InstanceCreateDetails,
     InstanceSpecificDetails,
     SessionDetails,
+    SessionError,
     TenantDetails,
     TimeParser,
     WaitResult,
@@ -70,7 +71,7 @@ def test_create_attached_session(requests_mock: Mocker) -> None:
         additional_matcher=assert_body,
     )
 
-    result = api.create_session(
+    result = api.get_or_create_session(
         name="name-0", dbid="dbid-1", pwd="pwd-2", memory=SessionMemory.m_4GB.value, ttl=timedelta(seconds=42)
     )
 
@@ -125,7 +126,7 @@ def test_create_dedicated_session(requests_mock: Mocker) -> None:
         additional_matcher=assert_body,
     )
 
-    result = api.create_session(
+    result = api.get_or_create_session(
         "name-0",
         "pwd-2",
         SessionMemory.m_4GB.value,
@@ -150,7 +151,7 @@ def test_create_dedicated_session(requests_mock: Mocker) -> None:
     )
 
 
-def test_create_standalone_session_error_forwards(requests_mock: Mocker) -> None:
+def test_create_standalone_session_http_error_forwards(requests_mock: Mocker) -> None:
     api = AuraApi(client_id="", client_secret="", tenant_id="some-tenant")
 
     mock_auth_token(requests_mock)
@@ -166,13 +167,70 @@ def test_create_standalone_session_error_forwards(requests_mock: Mocker) -> None
     )
 
     with pytest.raises(AuraApiError, match="some validation error"):
-        api.create_session(
+        api.get_or_create_session(
             "name-0",
             "pwd-2",
             SessionMemory.m_4GB.value,
             ttl=timedelta(seconds=42),
             cloud_location=CloudLocation("invalidProvider", "leipzig-1"),
         )
+
+
+def test_create_standalone_session_state_error_forwards(requests_mock: Mocker) -> None:
+    api = AuraApi(client_id="", client_secret="", tenant_id="some-tenant")
+
+    mock_auth_token(requests_mock)
+
+    requests_mock.post(
+        "https://api.neo4j.io/v1beta5/data-science/sessions",
+        status_code=200,
+        json={
+            "data": {
+                "id": "id0",
+                "name": "name-0",
+                "status": "Failed",
+                "instance_id": "",
+                "created_at": "1970-01-01T00:00:00Z",
+                "host": "1.2.3.4",
+                "memory": "4Gi",
+                "tenant_id": "tenant-0",
+                "user_id": "user-0",
+                "ttl": "42.0s",
+            },
+            "errors": [
+                {
+                    "id": "id0",
+                    "message": "Session reached its memory limit. Create a larger instance.",
+                    "reason": "OutOfMemory",
+                }
+            ],
+        },
+    )
+
+    session = api.get_or_create_session(
+        "name-0",
+        "pwd-2",
+        SessionMemory.m_4GB.value,
+        ttl=timedelta(seconds=42),
+        cloud_location=CloudLocation("gcp", "leipzig-1"),
+    )
+
+    session == SessionDetails(
+        id="id0",
+        name="name-0",
+        status="Failed",
+        instance_id="",
+        created_at=TimeParser.fromisoformat("1970-01-01T00:00:00Z"),
+        host="1.2.3.4",
+        memory=SessionMemory.m_4GB.value,
+        expiry_date=TimeParser.fromisoformat("1977-01-01T00:00:00Z"),
+        tenant_id="tenant-0",
+        user_id="user-0",
+        ttl=timedelta(seconds=42),
+        errors=[
+            SessionError(reason="OutOfMemory", message="Session reached its memory limit. Create a larger instance.")
+        ],
+    )
 
 
 def test_get_session(requests_mock: Mocker) -> None:
@@ -212,6 +270,57 @@ def test_get_session(requests_mock: Mocker) -> None:
         ttl=None,
         tenant_id="tenant-0",
         user_id="user-0",
+    )
+
+
+def test_get_session_state_error_forwards(requests_mock: Mocker) -> None:
+    api = AuraApi(client_id="", client_secret="", tenant_id="some-tenant")
+
+    mock_auth_token(requests_mock)
+
+    requests_mock.get(
+        "https://api.neo4j.io/v1beta5/data-science/sessions/id0",
+        json={
+            "data": {
+                "id": "id0",
+                "name": "name-0",
+                "status": "Failed",
+                "instance_id": "dbid-1",
+                "created_at": "1970-01-01T00:00:00Z",
+                "host": "1.2.3.4",
+                "memory": "4Gi",
+                "expiry_date": "1977-01-01T00:00:00Z",
+                "tenant_id": "tenant-0",
+                "user_id": "user-0",
+                "ttl": "42s",
+            },
+            "errors": [
+                {
+                    "id": "id0",
+                    "message": "Session reached its memory limit. Create a larger instance.",
+                    "reason": "OutOfMemory",
+                }
+            ],
+        },
+    )
+
+    session = api.get_session("id0")
+
+    assert session == SessionDetails(
+        id="id0",
+        name="name-0",
+        status="Failed",
+        instance_id="dbid-1",
+        created_at=TimeParser.fromisoformat("1970-01-01T00:00:00Z"),
+        host="1.2.3.4",
+        memory=SessionMemory.m_4GB.value,
+        expiry_date=TimeParser.fromisoformat("1977-01-01T00:00:00Z"),
+        tenant_id="tenant-0",
+        user_id="user-0",
+        ttl=timedelta(seconds=42),
+        errors=[
+            SessionError(reason="OutOfMemory", message="Session reached its memory limit. Create a larger instance.")
+        ],
     )
 
 
@@ -355,6 +464,88 @@ def test_list_sessions_with_db_id(requests_mock: Mocker) -> None:
     assert result == [expected1, expected2]
 
 
+def test_list_session_state_error_forwards(requests_mock: Mocker) -> None:
+    api = AuraApi(client_id="", client_secret="", tenant_id="some-tenant")
+
+    mock_auth_token(requests_mock)
+
+    requests_mock.get(
+        "https://api.neo4j.io/v1beta5/data-science/sessions",
+        json={
+            "data": [
+                {
+                    "id": "id0",
+                    "name": "name-0",
+                    "status": "Ready",
+                    "instance_id": "dbid-1",
+                    "created_at": "1970-01-01T00:00:00Z",
+                    "host": "1.2.3.4",
+                    "memory": "4Gi",
+                    "expiry_date": "1977-01-01T00:00:00Z",
+                    "tenant_id": "tenant-1",
+                    "user_id": "user-1",
+                },
+                {
+                    "id": "id1",
+                    "name": "name-2",
+                    "status": "Creating",
+                    "instance_id": "dbid-3",
+                    "created_at": "2012-01-01T00:00:00Z",
+                    "memory": "8Gi",
+                    "host": "foo.bar",
+                    "tenant_id": "tenant-2",
+                    "user_id": "user-2",
+                },
+                {
+                    "id": "id3",
+                    "name": "name-3",
+                    "status": "Creating",
+                    "instance_id": "dbid-3",
+                    "created_at": "2012-01-01T00:00:00Z",
+                    "memory": "8Gi",
+                    "host": "foo.bar",
+                    "tenant_id": "tenant-2",
+                    "user_id": "user-2",
+                },
+            ],
+            "errors": [
+                {
+                    "id": "id0",
+                    "message": "Session reached its memory limit. Create a larger instance.",
+                    "reason": "OutOfMemory",
+                },
+                {
+                    "id": "id1",
+                    "message": "Error 1.",
+                    "reason": "Reason1",
+                },
+                {
+                    "id": "id1",
+                    "message": "Error 2.",
+                    "reason": "Reason2",
+                },
+            ],
+        },
+    )
+
+    sessions = api.list_sessions()
+    errors = {s.id: s.errors for s in sessions}
+
+    assert errors["id0"] == [
+        SessionError(
+            message="Session reached its memory limit. Create a larger instance.",
+            reason="OutOfMemory",
+        )
+    ]
+
+    assert errors["id1"] == [
+        SessionError(message="Error 1.", reason="Reason1"),
+        SessionError(message="Error 2.", reason="Reason2"),
+    ]
+
+    assert errors["id3"] is None
+
+
 def test_delete_session(requests_mock: Mocker) -> None:
     api = AuraApi(client_id="", client_secret="", tenant_id="some-tenant")
 
@@ -453,6 +644,40 @@ def test_wait_for_session_running(requests_mock: Mocker) -> None:
     api = AuraApi("", "", tenant_id="some-tenant")
 
     assert api.wait_for_session_running("id0") == WaitResult.from_connection_url("neo4j+s://foo.bar")
+
+
+def test_wait_for_session_running_until_failure(requests_mock: Mocker) -> None:
+    mock_auth_token(requests_mock)
+    requests_mock.get(
+        "https://api.neo4j.io/v1beta5/data-science/sessions/id0",
+        json={
+            "data": {
+                "id": "id0",
+                "name": "name-0",
+                "status": "Failed",
+                "instance_id": "dbid-1",
+                "created_at": "1970-01-01T00:00:00Z",
+                "host": "foo.bar",
+                "memory": "4Gi",
+                "expiry_date": "1977-01-01T00:00:00Z",
+                "tenant_id": "tenant-1",
+                "user_id": "user-1",
+            },
+            "errors": [
+                {
+                    "id": "id0",
+                    "message": "Session reached its memory limit. Create a larger instance.",
+                    "reason": "OutOfMemory",
+                },
+            ],
+        },
+    )
+
+    api = AuraApi("", "", tenant_id="some-tenant")
+
+    assert api.wait_for_session_running("id0") == WaitResult.from_error(
+        "Session `id0` failed due to: [SessionError(message='Session reached its memory limit. Create a larger instance.', reason='OutOfMemory')]"
+    )
 
 
 def test_delete_instance(requests_mock: Mocker) -> None:
@@ -896,7 +1121,7 @@ def test_parse_session_info() -> None:
         "tenant_id": "tenant-1",
         "user_id": "user-1",
     }
-    session_info = SessionDetails.from_json(session_details)
+    session_info = SessionDetails.from_json(session_details, errors=[])
 
     assert session_info == SessionDetails(
         id="test_id",
@@ -925,7 +1150,7 @@ def test_parse_session_info_without_optionals() -> None:
         "tenant_id": "tenant-1",
         "user_id": "user-1",
     }
-    session_info = SessionDetails.from_json(session_details)
+    session_info = SessionDetails.from_json(session_details, errors=[])
 
     assert session_info == SessionDetails(
         id="test_id",