+++ changelog.md
@@ -14,5 +14,7 @@
 
 ## Improvements
 
+* Inform about Session failures such as `Out of Memory`. `GdsSession::get_or_create` will fail and `GdsSessions::list_sessions` will return the errors in a new field along `errors`.
+
 
 ## Other changes